app/requirements.txt
fastapi
uvicorn
jinja2
python-dotenv
httpx
supabase
python-multipart

